+++ src/main/kotlin/me/rafaelldi/aspire/actions/dashboard/StopResourceAction.kt
@@ -3,6 +3,8 @@ package me.rafaelldi.aspire.actions.dashboard
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
+import kotlinx.coroutines.launch
+import me.rafaelldi.aspire.AspireService
 import me.rafaelldi.aspire.generated.ResourceState
 import me.rafaelldi.aspire.generated.ResourceType
 import me.rafaelldi.aspire.sessionHost.SessionManager
@@ -21,7 +23,9 @@ class StopResourceAction : AnAction() {
             return
         }
 
-        SessionManager.getInstance(project).stopResource(resourceUid)
+        AspireService.getInstance(project).scope.launch {
+            SessionManager.getInstance(project).stopResource(resourceUid)
+        }
     }
 
     override fun update(event: AnActionEvent) {