+++ src/main/kotlin/me/rafaelldi/aspire/services/components/ResourceDashboardPanel.kt
@@ -1,6 +1,5 @@
 package me.rafaelldi.aspire.services.components
 
-import com.intellij.icons.AllIcons
 import com.intellij.ide.BrowserUtil
 import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.openapi.ui.DialogPanel
@@ -66,6 +65,10 @@ class ResourceDashboardPanel(resourceService: AspireResourceService) : BorderLay
             }
 
             if (resourceData.type == ResourceType.Project) {
+                val restartAction = ActionManager.getInstance().getAction("Aspire.Resource.Restart")
+                actionButton(restartAction)
+                val restartDebugAction = ActionManager.getInstance().getAction("Aspire.Resource.Restart.Debug")
+                actionButton(restartDebugAction)
                 val stopAction = ActionManager.getInstance().getAction("Aspire.Resource.Stop")
                 actionButton(stopAction)
             }